lesson2/seminar/step4_ge_validation_data/pyproject.toml
[tool.poetry]
name = "lesson2"
version = "0.1.0"
description = "Poetry dependencies"
authors = ["MLOps Student <student@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
pandas = ">=2.2,<3.0"
scikit-learn = ">=1.5,<2.0"
dvc = ">=3.50,<4.0"
great-expectations = ">=0.18,<1.0"
pyyaml = ">=6.0,<7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pre-commit (>=4.3.0,<5.0.0)"
]
